+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * New command to move in a cardinal direction while navigating around terrain.
 * Small update to mailbox menu.
+* Move catch fish command to fishing menu.
 
 ## 0.0.3 (May 4, 2021)
 * Skip cutscene command